initial fork creating awbot-logr
The main difference from awbot is that awlogr is a server which logs all
activity in a specific channel. It is managed by systemd.

socat is used to interface with OpenSSL rather than calling openssl directly.
It is a little more robust.

Makefile has the addition of BINDIR and MANDIR. See INSTALL for details.

This is a complete fork. Changes to awlogr will likely not be applied to
awbot. It is expected that awlogr will eventually be replaced by the haskell
branch of awbot.