fork creating awbot-greeter
awbot-greeter is a welcome bot. See the man page for details.

The program structure is very similar to awbot-logr. There are a few changes
which could find their way back to other versions of awbot.

socat is called with the -T option. This may fix the problem of socat failing
to exit when the connection breaks.

The option nick is replaced with nicks, an array. This gives more choices for
alternative nicks.